Reducer refactor & listing version lists
It was decided to refactor the store to store full version objects, as
opposed to just version IDs - this was in order to facilitate keeping
items pinned to the top of the list, not only in the new "Select another
version to compare to" table, but also in the comparison detail view -
previously selecting a version, changing page then selecting a
comparison would only show one of the two selected versions when
comparing because the version list only stores a page at a time. This
also results in an incorrect comparison (everything just appears as
removed). This refactor work allows us to circumvent those issues
without caching every version of a record in the user agent.