Enable inter-page operablility for verison list
Previously the History Viewer would baulk at attempting to browse
between pages or compare versions that resided on different pages in the
version list to each other. Previously an attempt was made to fix this
by storing full version objects in the store, which worked to a point
(the selection list remained functional when in compare mode), however
comparing two versions still had bugs from the refactor. These have now
been addressed along with a few other tidy ups of either things noticed
while debugging, or as the result of peer review.